+++ sentry-samples/sentry-samples-android/src/main/java/io/sentry/samples/android/MyApplication.java
@@ -2,12 +2,14 @@ package io.sentry.samples.android;
 
 import android.app.Application;
 import android.os.StrictMode;
+import io.sentry.Sentry;
 
 /** Apps. main Application. */
 public class MyApplication extends Application {
 
   @Override
   public void onCreate() {
+    Sentry.startProfileSession();
     strictMode();
     super.onCreate();