+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -415,10 +415,12 @@ public final class Sentry {
               try {
                 // We always delete the config file for app start profiling
                 FileUtils.deleteRecursively(appStartProfilingConfigFile);
-                if (!options.isEnableAppStartProfiling()) {
+                if (!options.isEnableAppStartProfiling() && !options.isStartProfilerOnAppStart()) {
                   return;
                 }
-                if (!options.isTracingEnabled()) {
+                // isStartProfilerOnAppStart doesn't need tracing, as it can be started/stopped
+                // manually
+                if (!options.isStartProfilerOnAppStart() && !options.isTracingEnabled()) {
                   options
                       .getLogger()
                       .log(
@@ -427,8 +429,13 @@ public final class Sentry {
                   return;
                 }
                 if (appStartProfilingConfigFile.createNewFile()) {
+                  // If old app start profiling is false, it means the transaction will not be
+                  // sampled, but we create the file anyway to allow continuous profiling on app
+                  // start
                   final @NotNull TracesSamplingDecision appStartSamplingDecision =
-                      sampleAppStartProfiling(options);
+                      options.isEnableAppStartProfiling()
+                          ? sampleAppStartProfiling(options)
+                          : new TracesSamplingDecision(false);
                   final @NotNull SentryAppStartProfilingOptions appStartProfilingOptions =
                       new SentryAppStartProfilingOptions(options, appStartSamplingDecision);
                   try (final OutputStream outputStream =
@@ -566,7 +573,8 @@ public final class Sentry {
     }
 
     final String profilingTracesDirPath = options.getProfilingTracesDirPath();
-    if (options.isProfilingEnabled() && profilingTracesDirPath != null) {
+    if ((options.isProfilingEnabled() || options.isContinuousProfilingEnabled())
+        && profilingTracesDirPath != null) {
 
       final File profilingTracesDir = new File(profilingTracesDirPath);
       profilingTracesDir.mkdirs();
@@ -1102,6 +1110,18 @@ public final class Sentry {
       final @NotNull TransactionContext transactionContext,
       final @NotNull TransactionOptions transactionOptions) {
     return getCurrentScopes().startTransaction(transactionContext, transactionOptions);
+  }
+
+  /** Starts the continuous profiler, if enabled. */
+  @ApiStatus.Experimental
+  public static void startProfileSession() {
+    getCurrentScopes().startProfileSession();
+  }
+
+  /** Stops the continuous profiler, if enabled. */
+  @ApiStatus.Experimental
+  public static void stopProfileSession() {
+    getCurrentScopes().stopProfileSession();
   }
 
   /**